requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
requests==2.31.0
mcp>=0.9.0
pytz>=2023.3
